# Improved colliderect() function