## Info
Follows #3324. One of #3337
The old colliderect implementation relied heavily on automatic argument "grouping" as a mean to implicitly construct the rect if more than 1 argument were passed. 
This new implementation relies upon direct access to the elements via FASTCALL, enabling faster argument handling and a more direct collision detection pipeline.
I've also added another collision detection function for normalised rects. Using that speeds up the current implementation by an additional 5-6%.

## Performance Stats
Overall the new implementation is 36.98% faster than the current one, but in the case of a single Rect argument(69.68% better) or 4 arguments can get up to 118% faster.
Using CPython 3.6 or older i saw a 7.1% mean performance reduction(EDIT. now is up to par with 2.1.4 values)
Using the code below i got the following results:

**--- OLD---**
```
--------- Not normalized Colliding rects ---------
Rect:  0.6726158000019495
4 args:  1.2384678999987955
2 args:  1.3053118999996514
1 tuple arg:  1.1832979999999225

--------- Normalized Colliding rects ---------
Rect:  0.6773214999993797
4 args:  1.2457533000015246
2 args:  1.309086199999001
1 tuple arg:  1.1767221999980393

--------- Non-colliding rects ---------
Rect:  0.6247191000002204
4 args:  1.2045317999982217
2 args:  1.29535400000168
1 tuple arg:  1.1573203999978432
13.090502099996229  seconds
```

**--- NEW ---**
```
--------- Not normalized Colliding rects ---------
Rect:  0.41285819999757223
4 args:  0.9083898999997473
2 args:  1.1966508999976213
1 tuple arg:  1.0859782999978052

--------- Normalized Colliding rects ---------
Rect:  0.4182092000010016
4 args:  0.6948287000013806
2 args:  1.031187399999908
1 tuple arg:  1.0486275000002934

--------- Non-colliding rects ---------
Rect:  0.3959212999980082
4 args:  0.7050670999997237
2 args:  1.0309612000019115
1 tuple arg:  1.0551780999994662
9.98385779999444  seconds
```
**--- PYTHON 3.6.7---**
```
--------- Not normalized Colliding rects ---------
Rect:  0.7407949
4 args:  1.2265097
2 args:  1.3928916
1 tuple arg:  1.3093952

--------- Normalized Colliding rects ---------
Rect:  0.7273856999999992
4 args:  1.1870614000000002
2 args:  1.2928631
1 tuple arg:  1.3021802999999998

--------- Non-colliding rects ---------
Rect:  0.7285181999999999
4 args:  1.1909159000000002
2 args:  1.2929525999999996
1 tuple arg:  1.344197600000001
13.7356662  seconds
```

I used this code:
```Python
import timeit

import pygame

pygame.init()
NUMBER = 10000000
r1 = pygame.rect.Rect(0, 0, 10, 10)
r2 = pygame.rect.Rect(5, 5, 10, 10)
r3 = pygame.rect.Rect(15, 15, 10, 10)

nnr1 = pygame.rect.Rect(10, 10, -10, -10)
nnr2 = pygame.rect.Rect(15, 15, -10, -10)
tottime = 0
GLOB = {"r1": r1, "r2": r2, "r3": r3, "nnr1": nnr1, "nnr2": nnr2}

print("\n--------- Not normalized Colliding rects ---------")

timed = timeit.timeit("nnr1.colliderect(nnr2)",
                      globals=GLOB, number=NUMBER)
print("Rect: ", timed)
tottime += timed

timed = timeit.timeit("nnr1.colliderect(15, 15, -10, -10)",
                      globals=GLOB, number=NUMBER)
print("4 args: ", timed)
tottime += timed

timed = timeit.timeit("nnr1.colliderect((15, 15), (-10, -10))",
                      globals=GLOB, number=NUMBER)
print("2 args: ", timed)
tottime += timed

timed = timeit.timeit("nnr1.colliderect((15, 15, -10, -10))",
                      globals=GLOB, number=NUMBER)
print("1 tuple arg: ", timed)
tottime += timed

print("\n--------- Normalized Colliding rects ---------")
timed = timeit.timeit("r1.colliderect(r2)",
                      globals=GLOB, number=NUMBER)
print("Rect: ", timed)
tottime += timed

timed = timeit.timeit("r1.colliderect(5, 5, 10, 10)",
                      globals=GLOB, number=NUMBER)
print("4 args: ", timed)
tottime += timed

timed = timeit.timeit("r1.colliderect((5, 5), (10, 10))",
                      globals=GLOB, number=NUMBER)
print("2 args: ", timed)
tottime += timed

timed = timeit.timeit("r1.colliderect((5, 5, 10, 10))",
                      globals=GLOB, number=NUMBER)
print("1 tuple arg: ", timed)
tottime += timed

print("\n--------- Non-colliding rects ---------")

timed = timeit.timeit("r1.colliderect(r3)",
                      globals=GLOB, number=NUMBER)
print("Rect: ", timed)
tottime += timed

timed = timeit.timeit("r1.colliderect(15, 15, 10, 10)",
                      globals=GLOB, number=NUMBER)
print("4 args: ", timed)
tottime += timed

timed = timeit.timeit("r1.colliderect((15, 15), (10, 10))",
                      globals=GLOB, number=NUMBER)
print("2 args: ", timed)
tottime += timed

timed = timeit.timeit("r1.colliderect((15, 15, 10, 10))",
                      globals=GLOB, number=NUMBER)
print("1 tuple arg: ", timed)
tottime += timed

print(tottime, " seconds")
```