@itzpr3d4t0r 2022-07-19
New performance metrics with the added _pg_do_notnormalized_rects_intersect() function : 
(notice how the 4 args not normalized case went from 0.9083898999997473s down to 0.742086800000834s)
```
--------- Not normalized Colliding rects ---------
Rect:  0.40322829999786336
4 args:  0.742086800000834
2 args:  1.1501164000001154
1 tuple arg:  1.025158200001897

--------- Normalized Colliding rects ---------
Rect:  0.40222820000053616
4 args:  0.6949547999975039
2 args:  1.0398936999990838
1 tuple arg:  1.0740198000021337

--------- Non-colliding rects ---------
Rect:  0.40684569999939413
4 args:  0.7239664000007906
2 args:  1.057071499999438
1 tuple arg:  1.0149845000014466
9.734554300001037  seconds
```