@itzpr3d4t0r 2022-07-20
New stats after latest commit:

```
--------- Not normalized Colliding rects ---------
Rect:  0.40994370000044
4 args:  0.6392197999994096
2 args:  1.141009800001484
1 tuple arg:  1.0302646000018285

--------- Normalized Colliding rects ---------
Rect:  0.3991593999999168
4 args:  0.5655972999993537
2 args:  1.0292699000019638
1 tuple arg:  1.0392322999978205

--------- Non-colliding rects ---------
Rect:  0.3933633999986341
4 args:  0.5685177999985171
2 args:  1.0291066999998293
1 tuple arg:  1.0239384999986214
9.268623199997819  seconds

```