@itzpr3d4t0r 2022-07-20
Latest test
```
--------- Not normalized Colliding rects ---------

Rect:  0.40383030000157305
4 args:  0.7056063999989419
2 args:  1.1588473999981943
1 tuple arg:  1.0563916999999492
--------- Normalized Colliding rects ---------

Rect:  0.40069299999959185
4 args:  0.669831899998826
2 args:  1.0328869999975723
1 tuple arg:  1.0233136999995622
--------- Non-colliding rects ---------

Rect:  0.39414380000016536
4 args:  0.6520156000005954
2 args:  1.0341171000000031
1 tuple arg:  1.0242848000016238
9.555962699996599  seconds
```