@itzpr3d4t0r 2022-07-25
Latest test:
```
--------- Not normalized Colliding rects ---------
Rect:  0.38273279999884835
4 args:  0.6658422000000428
2 args:  1.1407941999987088
1 tuple arg:  0.9429583000001003

--------- Normalized Colliding rects ---------
Rect:  0.4201164999994944
4 args:  0.6050114999998186
2 args:  1.0299127000016597
1 tuple arg:  0.9367297000007966

--------- Non-colliding rects ---------
Rect:  0.39076250000107393
4 args:  0.6168821000010212
2 args:  1.0340894999990269
1 tuple arg:  0.9273178000003099
9.093149800000901  seconds
```